fix: use of `LocalizedStringKey` in `Toggle`, `Button`, `Action` (#2229)
The current implementation of the Toggle, Button and Label in extensions relis on the key of the LocalizedStringResource
or the String which were not converted as localizable.
By converting this String to a LocalizedStringKey, in case of a String parameter, we ensure that:
- given an "Account" text it will load the "Account"-keyed wording, i.e. the suitable translated wording
- given a "Foo" text without translations the "Foo" text will be displayed.

This bug occured when using another language as english.
Otherwise because wording keys are plain english words it is more tricky to spot.

Signed-off-by: Pierre-Yves Lapersonne <dev@pylapersonne.info>